+++ lib/telegem.rb
@@ -3,7 +3,7 @@ require 'logger'
 require 'json'
 
 module Telegem
-  VERSION = "3.4.0"
+  VERSION = "3.5.0"
 end
 
 #